// src/types/master/ItemTypes.d.ts
export interface ItemFormData {
  categoryId: string;
  modelId: string;
  sizeId: string;
  colourId: string;
  itemName: string;
  itemShortName: string;
  unitId: string;
  purchaseLedgerId: string;
  salesLedgerId: string;
  cgst: string;
  sgst: string;
  igst: string;
  purchaseRate: string;
  salesRate: string;
  openingQuantity: string;
  openingRate: string;
  total: string;
}
// ...
export interface ItemTableData {
  Id: number;
  Cat_Id: numberl;
  Model_Id: number | null;
  Size_Id: number | null;
  Color_Id: number | null;
  Item_Name: string;
  Item_Sh_Name: string;
  Unit_Id: number;
  Purchase_Gl: number;
  Sales_Gl: number;
  CGST: string;
  SGST: string;
  IGST: string;
  Pur_Rate: string;
  Sale_Rate: string;
  Cat_Name: string;
  Open_Qnty: string;
  Item_Rate: string;
}